% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robkat_utils.R
\name{bisquarePsiF}
\alias{bisquarePsiF}
\title{Tukey's Bi-square psi function}
\usage{
bisquarePsiF(u, k = 4.685)
}
\arguments{
\item{\code{u}}{An nx1 vector}

\item{\code{k}}{Cut-off parameter}
}
\description{
Tukey's Bi-square psi function
}
\details{
Tukey's Bi-square psi function is defined as
          psi(u) = u * (1-(u/k)^2)^2 * (abs(u)<=k)
}
\examples{
u <- seq(-10, 10, len = 101)
plot(u, bisquarePsiF(u), type="b")
}
