% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robkat_utils.R
\name{check.Z.MinorAllels}
\alias{check.Z.MinorAllels}
\title{Function for checking whether minor allele frequencies of SNPs are
                correctly supplied}
\usage{
check.Z.MinorAllels(Z, silent = FALSE)
}
\arguments{
\item{\code{Z}}{An nxp matrix of genetic covariates to be tested. Each column
corresponds to one SNP.}

\item{\code{silent}}{Whether to print a warning if genotypes are flipped}
}
\description{
Function for checking whether minor allele frequencies of SNPs are
                correctly supplied
}
\details{
For each column of \code{Z}, we compute [2*(# of 2's) +
             (# of 1's)] and  [2*(# of 0's) + (# of 1's)]. If the former
             is smaller than the latter quantity, the MAF is correct. Otherwise,
             genotypes of that column are exchanged (0 -> 2 and 2 -> 0)
}
\examples{
n <- 100 #sample size
p <- 5 #no. of SNPs variables
Z <- matrix(sample(0:2, n*p, replace = T), ncol=p) # pretending to be SNP data
Znew <- check.Z.MinorAllels(Z, silent = FALSE)
}
