% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robkat_utils.R
\name{hampelPsiF}
\alias{hampelPsiF}
\title{Hampel's psi function}
\usage{
hampelPsiF(u, a = 1.353, b = 3.157, r = 7.216)
}
\arguments{
\item{\code{u}}{An nx1 vector}

\item{\code{k}}{Cut-off parameter}
}
\description{
Hampel's psi function
}
\details{
Tukey's Bi-square psi function is defined as
          psi(u) = u*(abs(u)<=a) +
          a*sign(u)*((abs(u) > a) & (abs(u) <=b)) +
          a*sign(u)*((r-abs(u))/(r-b))*((abs(u) > b) & (abs(u) <=r))
          
          The default values of a=1.353, b=3.157, and r=7.216 are 
          chosen for 95 \% efficiency compared to the least squares 
          estimate when the error distribution is normal.
}
\examples{
u <- seq(-10, 10, len = 101)
plot(u, hampelPsiF(u), type="b")
}
