% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robkat_utils.R
\name{huberPsiF}
\alias{huberPsiF}
\title{Huber's psi function}
\usage{
huberPsiF(u, k = 1.345)
}
\arguments{
\item{\code{u}}{An nx1 vector}

\item{\code{k}}{Cut-off parameter}
}
\description{
Huber's psi function
}
\details{
Huber loss behaves like least squares loss until a cut-off point, k,
          after which the loss behaves like absolute value loss. 
          The default value, k = 1.345, is chosen for 95 \% efficiency 
          compared to the least squares estimate when the error distribution is normal. 
          Huber's psi function is defined as
          psi(u) = u*[abs(u) <= k] + k*sign(u)*[abs(u) > k]. It 
          is the derivative of rho(x) = 0.5*x^2*[abs(x)<= k] + k*(abs(x)-0.5*k)*[abs(x) > k]
}
\examples{
u <- seq(-10, 10, len = 101)
plot(u, huberPsiF(u), type="b")
}
