% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robkat_utils.R
\name{LADPsiF}
\alias{LADPsiF}
\title{LAD Check function}
\usage{
LADPsiF(u)
}
\arguments{
\item{\code{u}}{An nx1 vector}

\item{\code{k}}{Cut-off parameter}
}
\description{
LAD Check function
}
\details{
The check function for LAD is defined as
        psi(u) = sign(u)/2.
        If u  is zero, then psi(u) is assigned either -1/2 or 1/2 with
        probability 1/2. LAD loss represents the subgradient of rho(x) = abs(x).
}
\examples{
u <- seq(-10, 10, len = 101)
plot(u, LADPsiF(u), type="b")
}
