% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_funs.R
\name{linearKernel}
\alias{linearKernel}
\title{Compute the linear kernel}
\usage{
linearKernel(Z)
}
\arguments{
\item{\code{Z}}{A SNP matrix of size \code{n x p}. Here \code{n} is the
number of subjects and \code{p} is the number of SNPs. Each
entry of \code{Z} should be 0, 1 or 2 denoting the minor
allele frequency.}
}
\description{
Compute the linear kernel
}
