% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_funs.R
\name{quadraticKernel}
\alias{quadraticKernel}
\title{Compute the quadratic kernel}
\usage{
quadraticKernel(Z)
}
\arguments{
\item{\code{Z}}{A SNP matrix of size \code{n x M}. Here \code{n} is the
number of subjects and \code{M} is the number of SNPs. Each
entry of \code{Z} should be 0, 1 or 2 denoting the minor
allele frequency.}
}
\description{
Compute the quadratic kernel
}
