% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robkat.R
\name{robkat}
\alias{robkat}
\title{Robust Kernel Association Test}
\usage{
robkat(Y, Z, X = NULL, kernel.fun = "linear", loss.fun = "huber",
  silent = FALSE)
}
\arguments{
\item{\code{Y}}{An nx1 response vector}

\item{\code{Z}}{An nxM matrix of genetic covariates to be tested}

\item{\code{X}}{An nxp matrix of non-genetic covariates, if present.
If no other covariates, set X = NULL. Do not
include an intercept column.}

\item{\code{kernel.fun}}{A string ("linear", "quadratic", "IBS")
indicating which kernel function to use}

\item{\code{loss.fun}}{A string ("huber", "hampel", "bisquare")
indicating which loss function to use}

\item{\code{silent}}{Whether to print a warning if genotypes are
flipped when the MAFs are checked for validity.}
}
\description{
Robust Kernel Association Test
}
\examples{
n <- 100 #sample size
p <- 5 #no. of SNPs variables
q <- 2 #no. of parametric variables
bet = rep(1, q) #params for parametric effects
Z <- matrix(sample(0:2, n*p, replace = T), ncol=p) # pretending to be SNP data
X <- matrix(rnorm(n*q), n, q) # some other covariates
Y <- X\%*\%bet + 0.5*rowSums(Z) + rcauchy(n)
robkat(Y, Z, X, kernel.fun = "IBS", loss.fun = "huber")
robkat(Y, Z, X, kernel.fun = "IBS", loss.fun = "LAD")
}
\keyword{Robust}
